feat: Merge all reporters to one, create through factory
