+++ core/src/reporter/data-feed.ts
@@ -1,94 +1,19 @@
-import { Job, Worker } from 'bullmq'
 import { Logger } from 'pino'
 import type { RedisClientType } from 'redis'
-import { State } from './state'
-import { sendTransaction } from './utils'
-import { watchman } from './watchman'
+import { factory } from './factory'
 import {
   REPORTER_AGGREGATOR_QUEUE_NAME,
-  BULLMQ_CONNECTION,
-  CHAIN,
   DATA_FEED_REPORTER_STATE_NAME,
-  DATA_FEED_SERVICE_NAME,
-  PROVIDER_URL
+  DATA_FEED_SERVICE_NAME
 } from '../settings'
-import { ITransactionParameters } from '../types'
-import { OraklError, OraklErrorCode } from '../errors'
 
-const FILE_NAME = import.meta.url
-
-export async function reporter(redisClient: RedisClientType, _logger: Logger) {
-  const logger = _logger.child({ name: 'reporter', file: FILE_NAME })
-
-  const state = new State({
+export async function buildReporter(redisClient: RedisClientType, logger: Logger) {
+  await factory({
     redisClient,
-    providerUrl: PROVIDER_URL,
     stateName: DATA_FEED_REPORTER_STATE_NAME,
     service: DATA_FEED_SERVICE_NAME,
-    chain: CHAIN,
-    logger
+    reporterQueueName: REPORTER_AGGREGATOR_QUEUE_NAME,
+    concurrency: 5,
+    _logger: logger
   })
-  await state.refresh()
-
-  logger.debug(await state.active(), 'Active reporters')
-
-  const DATA_FEED_REPORTER_WORKER_CONCURRENCY = 5
-  const reporterWorker = new Worker(REPORTER_AGGREGATOR_QUEUE_NAME, await job(state, logger), {
-    ...BULLMQ_CONNECTION,
-    concurrency: DATA_FEED_REPORTER_WORKER_CONCURRENCY
-  })
-  reporterWorker.on('error', (e) => {
-    logger.error(e)
-  })
-
-  const watchmanServer = await watchman({ state, logger })
-
-  async function handleExit() {
-    logger.info('Exiting. Wait for graceful shutdown.')
-
-    await redisClient.quit()
-    await reporterWorker.close()
-    await watchmanServer.close()
-  }
-  process.on('SIGINT', handleExit)
-  process.on('SIGTERM', handleExit)
-}
-
-function job(state: State, logger: Logger) {
-  async function wrapper(job: Job) {
-    const inData: ITransactionParameters = job.data
-    logger.debug(inData, 'inData')
-
-    const { payload, gasLimit, to } = inData
-
-    const wallet = state.wallets[to]
-    if (!wallet) {
-      const msg = `Wallet for oracle ${to} is not active`
-      logger.error(msg)
-      throw new OraklError(OraklErrorCode.WalletNotActive, msg)
-    }
-
-    const NUM_TRANSACTION_TRIALS = 3
-    for (let i = 0; i < NUM_TRANSACTION_TRIALS; ++i) {
-      try {
-        await sendTransaction({ wallet, to, payload, logger, gasLimit })
-        break
-      } catch (e) {
-        if (
-          ![
-            OraklErrorCode.TxNotMined,
-            OraklErrorCode.TxProcessingResponseError,
-            OraklErrorCode.TxMissingResponseError
-          ].includes(e.code)
-        ) {
-          throw e
-        }
-
-        logger.info(`Retrying transaction. Trial number: ${i}`)
-      }
-    }
-  }
-
-  logger.debug('Reporter job built')
-  return wrapper
 }
